Remove navegação da callback onSuccess
