+++ src/queries/user.ts
@@ -1,5 +1,4 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { useLocation, useNavigate } from "react-router-dom";
 
 import { resetAllStores } from "@/store/helpers";
 import { saveTokens } from "@/store/tokens";
@@ -9,10 +8,6 @@ import * as userApi from "../services/api/user";
 export const useUserQueries = () => {
   const queryClient = useQueryClient();
 
-  const navigate = useNavigate();
-
-  const location = useLocation();
-
   const useGetUser = (enabled: boolean) =>
     useQuery({
       queryKey: ["user"],
@@ -20,28 +15,18 @@ export const useUserQueries = () => {
       enabled,
     });
 
+  const fetchUser = (accessToken: string) =>
+    queryClient.fetchQuery({
+      queryKey: ["user"],
+      queryFn: () =>
+        userApi.getUser(accessToken).then((response) => response.data),
+    });
+
   const useAuthUser = () =>
     useMutation({
       mutationFn: userApi.authenticateUser,
-      onSuccess: async (tokens) => {
-        const user = await queryClient.fetchQuery({
-          queryKey: ["user"],
-          queryFn: () =>
-            userApi
-              .getUser(tokens.accessToken)
-              .then((response) => response.data),
-        });
-
+      onSuccess: (tokens) => {
         saveTokens(tokens);
-
-        const from =
-          location.state?.from?.pathname || user.tipo === "aluno"
-            ? "/perfil-aluno"
-            : user.tipo === "professor"
-              ? "/perfil-professor"
-              : "/perfil-coordenador";
-
-        navigate(from, { replace: true });
       },
     });
 
@@ -62,6 +47,7 @@ export const useUserQueries = () => {
 
   return {
     useGetUser,
+    fetchUser,
     useAuthUser,
     useCreateAluno,
     useCreateProfessor,
